move generateTransportData() data from model.Species to species.Species
the transportData attribute was already part of species.Species,
so there was no reason to keep it in rmgpy.rmg.model.Species.
