+++ rmgpy/rmg/model.py
@@ -41,7 +41,6 @@ from rmgpy.display import display
 #import rmgpy.chemkin
 import rmgpy.constants as constants
 from rmgpy.constraints import failsSpeciesConstraints
-from rmgpy.data.rmg import getDB
 from rmgpy.quantity import Quantity
 import rmgpy.species
 from rmgpy.thermo import Wilhoit, NASA, ThermoData
@@ -98,21 +97,6 @@ class Species(rmgpy.species.Species):
         self.conformer.E0 = self.getThermoData().E0
         self.conformer.modes = conformer.modes
         self.conformer.spinMultiplicity = conformer.spinMultiplicity
-            
-    def generateTransportData(self):
-        """
-        Generate the transportData parameters for the species.
-        """
-
-        try:
-            transportDB = getDB('transport')        
-            if not transportDB: raise Exception
-        except Exception, e:
-            logging.debug('Could not obtain the transport database. Not generating transport...')
-            raise e
-
-        #count = sum([1 for atom in self.molecule[0].vertices if atom.isNonHydrogen()])
-        self.transportData = transportDB.getTransportProperties(self)[0]
     
     def generateEnergyTransferModel(self):
         """
